requirements.txt
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Deprecated==1.2.14
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gspread==5.12.3
httplib2==0.22.0
idna==3.6
lxml==4.9.4
markdown-it-py==3.0.0
mdurl==0.1.2
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
pyshorteners==1.0.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
urllib3==2.1.0
wrapt==1.16.0
